Fast speed limitation
'Resultant' is velocity vector's x and velocity vector's z resultant.
So 'resultant' is a speed via velocity. (estimated value)
'SpeedTime' is a "The time the current speed is driving"
When the speed reaches the limit, the timer will be activated and will not be accelerated during the timer.
If the acceleration stops(release shift) , the 'speedTime' will reverse.